a bit more refactoring to clean up files
